feat(util): moving weighted median
Cython implementation for 1D and 2D. Uses a (C++) binary tree to find
the medians fast and a queue for tracking the elements in the tree
and fast removal.

Uses OpenMP (cython prange) to parallelize over columns in 2-dimensional
data.

Accepts method option 'split', 'lower' and 'higher' that decide what's
the median in case there are multiple values that satisfy the median
conditions.