feat(BasicCont): make history an h5 attribute
add_history previously accepted single-level dictionaries only and they
were added as attributes to the history group. This was a simply
solution to dictionary de-/serialization with json). Now that we support
full de-/serialization of dictionaries using json, the history feature
was broken. To make it usable again, the history is now not a group
anymore, but an attribute, so that dictionaries of any depth can be
added.

Reading the old history format is still supported but creates a
deprecation warning.