fix(misc): change lock_file() behavior to match docstring
The lock_file() docstring states that only rank=0 will manipulate the file
if comm is specified, but in fact the routine checks mpiutil.rank0 if
comm is not specified. This commit changes this behavior to match what
the docstring says (which is important if multiple tasks each want to use
lock_file() for their own files).