fix(time): allow zero length arrays as arguments
This fixes many of the routines in caput.time to allow zero length
arrays as arguments. To do this it introduces a decorator `scalarize`
which allows code to be written which assumes it is receiving a non-zero
length array as an argument, while the decorator takes care of scalar
and zero length array arguments.