fix(time): Make Star.names iterable in skyfield_star_from_ra_dec
The skyfield documentation is poor, but I think the intention is
that the `Star.names` attribute is meant to be iterable.  The
skyfield code doesn't reference `Star.names` at all, and there's
no documentation on the meaning of the attribute (or any of the
parameters of `Star.__init__`), so it's very much unclear.

The default value for `Star.names` is an empty tuple.  Probably
any (non-`str`) iterable would be reasonable.

This changes `skyfield_star_from_ra_dec` to convert a `str` passed
as `name` into a single-element tuple when assigning it to
`Star.names`, and also updates the docstring to indicate that a
tuple or list of names is an acceptable value for the `name`
parameter.